Return empty list for no results from Solr
Change error response to return an empty list instead of an error message.

This is not really an error, we got no results from solr, which is a valid response.  So I think it's better we return an empty list here.  

Also the 404 is not the best response code here since it sometimes requires special handling to determine if it's a 404 because the service was not found or if it was this error.